# ouranos-war/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "0.0.0-SNAPSHOT",
    "title" : "Ouranos Swagger Maven Plugin Sample"
  },
  "host" : "localhost:8080",
  "basePath" : "/api",
  "tags" : [ {
    "name" : "visibleSat",
    "description" : "Operations about visible satellite"
  } ],
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/visibleSat/{version}/byPeriod" : {
      "get" : {
        "tags" : [ "visibleSat" ],
        "summary" : "Finds visible satellite order by period",
        "description" : "",
        "operationId" : "getVisiblesatelliteByPeriodJson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "version of the api",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "requete",
          "in" : "query",
          "description" : "submission request on json",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Given visible satellite found on json"
          }
        }
      },
      "post" : {
        "tags" : [ "visibleSat" ],
        "summary" : "Finds visible satellite order by period",
        "description" : "",
        "operationId" : "postVisiblesatelliteByPeriodJson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "version of the api",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "submission request on json",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Given visible satellite found"
          }
        }
      }
    },
    "/visibleSat/{version}/bySatellite" : {
      "get" : {
        "tags" : [ "visibleSat" ],
        "summary" : "Finds visible satellite order by satellite id",
        "description" : "",
        "operationId" : "getVisiblesatelliteBySatelliteJson",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "version of the api",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "submission request on json",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Given visible satellite found"
          }
        }
      },
      "post" : {
        "tags" : [ "visibleSat" ],
        "summary" : "Finds visible satellite order by satellite id",
        "description" : "",
        "operationId" : "postVisiblesatelliteBySatelliteXml",
        "consumes" : [ "application/xml" ],
        "produces" : [ "application/xml;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "version of the api",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "submission request on xml",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Given visible satellite found"
          }
        }
      }
    }
  }
}